Moved main method out of Module Manager
