+++ src/main/java/edu/mines/acmX/exhibit/module_management/ModuleManager.java
@@ -76,96 +76,6 @@ public class ModuleManager {
 	static Logger logger = LogManager.getLogger(ModuleManager.class.getName());
 
 	/**
-	 * Main function for the ModuleManager framework. Creates an instance of
-	 * ModuleManager and runs it.
-	 */
-	public static void main(String[] args) {
-		CommandLineParser cl = new GnuParser();
-		CommandLine cmd;
-		Options opts = generateCLOptions();
-		try {
-			cmd = cl.parse(opts, args);
-
-			if (cmd.hasOption("help")) {
-				printHelp(opts);
-			} else {
-
-				if (cmd.hasOption("manifest")) {
-					ModuleManager.configure(cmd.getOptionValue("manifest"));
-				} else {
-					System.out
-							.println("A Module Manager Manifest is required to run the module manager" + 
-                                    "Please specify with the --manifest switch");
-					System.exit(1);
-				}
-				ModuleManager m;
-				m = ModuleManager.getInstance();
-				m.run();
-			}
-
-		} catch (ParseException e) {
-			printHelp(opts);
-			logger.error("Incorrect command line arguments");
-			e.printStackTrace();
-		} catch (ManifestLoadException e) {
-			logger.fatal("Could not load the module manager manifest");
-			e.printStackTrace();
-		} catch (ModuleLoadException e) {
-			logger.fatal("Could not load the default module");
-			e.printStackTrace();
-		} catch (HardwareManagerManifestException e) {
-			logger.fatal("Could not load the manfiest for the Hardware Manager");
-			e.printStackTrace();
-		} catch (BadDeviceFunctionalityRequestException e) {
-			logger.fatal("Default module depends on unknown device functionalities");
-			e.printStackTrace();
-		} catch (InvalidConfigurationFileException e) {
-			logger.fatal(e.getMessage());
-			e.printStackTrace();
-		}
-
-	}
-
-	private static void printHelp(Options opts) {
-		HelpFormatter formatter = new HelpFormatter();
-        formatter.setDescPadding(0);
-        String header =
-            "\n" + 
-            "Welcome to the Interface SDK!\n" +
-            "The Interface SDK provides an intelligent environment in which to run modules. " + 
-            "To build a module visit the github wiki at " + 
-            "https://github.com/ColoradoSchoolOfMines/interface_sdk/wiki " +
-            "\n" + 
-            "Also please find the source code at " + 
-            "https://github.com/ColoradoSchoolOfMines/interface_sdk " + 
-            "where you can find more detail on the open source project.";
-        String footer = 
-            "\n";
-		formatter.printHelp("java -jar [JARNAME]", header, opts, footer,
-				true);
-	}
-
-	private static Options generateCLOptions() {
-		Options options = new Options();
-		// options = optionsUsingIndividualAgruments(options);
-		options.addOption(optionsUsingManifest());
-		// options.addOption(openNiArguments());
-		options.addOption("h", "help", false, "\nPrint these helpful hints");
-		return options;
-	}
-
-	private static Option optionsUsingManifest() {
-		return OptionBuilder.withLongOpt("manifest")
-				.withDescription(
-                        "\nUse a custom module manager manifest file. " + 
-                        "The manifest must specify the default module to load, " +
-                        "the location of the modules folder and any configuration files " +
-                        "that are needed for the drivers you would like to use."
-                        )
-				.hasArg().withArgName("PATH").create();
-	}
-
-	/**
 	 * Singleton instance of ModuleManager This is volatile in order to be safe
 	 * with multiple threads
 	 */
